Collect TSR diagnostics from Dell hosts
Summary:
Added "nfs_share" variable to .rackopsrc.
It should default to "62.217.124.76:/srv/www/dell/TSR/".
This option is passed down to dell.py class.

Functions added:
-  _ssh
- _find_jid
- _confirm_job
- diagnostics

They are all specific to Dell hosts.

Ran by "rackops diagnostics <idrac_serial>", it creates two ssh connections to
the ipmi, one to collect the TSR and one to export it to the forementioned
nfs share.

Between the "collect" and "export" commands we have a time.sleep() of
3 minutes.

This is required, because if the "collect" job does not complete the "export"
one will fail.

Also the collect job result changes between states,
(state 1 -> job is created/started, state 2 -> job is finished)
and we need to get the last result to find if it was actually successful.

The user has to have access to the nfs share to retrive the zip file.

Part of T1956.

Test Plan:
tested locally with a number of serials taken from the list in task
T1956.

Reviewers: #servers, vitsalis

Reviewed By: #servers, vitsalis

Subscribers: vitsalis, stefanosliv, nkorb, #servers

Differential Revision: https://phab.noc.grnet.gr/D5236